+++ changelog.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## 2022.04
 * Added: node and composer caching for GitHub actions
+* Updated: Webpack & related configs for React HMR / Example react app dev.
 * Fixed: deploy commit messages via GitHub actions now display the branch and environment instead of the variables.
 * Changed: all block models are now instantiated with the container to support dependency injection.